Forgot to say in last update that I fixed compiler warnings in EUTelClusteringProcessor.cc on lines 1013,1558,2551. Also now fixed compiler warnings in EUTelClusterFilter.cc on lines 819,893. All warnings were signed/unsigned mismatches

+++ src/EUTelClusterFilter.cc
@@ -9,7 +9,7 @@
  *   header with author names in all development based on this file.
  *
  */
-
+ 
 // eutelescope includes ".h"
 #include "EUTELESCOPE.h"
 #include "EUTelVirtualCluster.h"
@@ -816,7 +816,7 @@ void EUTelClusterFilter::processEvent (LCEvent * event) {
                                     // bool isHit  = ( statusMatrix->getADCValues()[index] ==
                                     // EUTELESCOPE::HITPIXEL );
                                     //
-                                    if( statusMatrix->getADCValues().size() > index )
+                                    if((int) statusMatrix->getADCValues().size() > index )
                                     {
                                     bool isBad  = ( statusMatrix->getADCValues()[index] == EUTELESCOPE::BADPIXEL );
                                     if ( !isBad )
@@ -890,7 +890,7 @@ void EUTelClusterFilter::processEvent (LCEvent * event) {
                                     // EUTELESCOPE::HITPIXEL );
                                     // 
                                     ///
-                                    if( statusMatrix->getADCValues().size() > index )
+                                    if((int) statusMatrix->getADCValues().size() > index )
                                     {
 
                                     bool isBad  = ( statusMatrix->getADCValues()[index] == EUTELESCOPE::BADPIXEL );
